Some configurations

+++ 0x04-more_functions_nested_loops/101-print_number.c
@@ -74,7 +74,7 @@ void print_number(int n)
 
 	for (i = count - 1; i >= 0; i--)
 	{
-		d = (tmp / (int) pow_int(10, i)) % 10;
+		d = (tmp / pow_int(10, i)) % 10;
 		_putchar(d + '0');
 	}
 }
